Unit test for Date.localizedDateFormatter() move swift logic to Date+Format.swift
