+++ Wire-iOS Tests/DateFormatterTests.swift
@@ -1,0 +1,45 @@
+//
+// Wire
+// Copyright (C) 2017 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+
+import XCTest
+@testable import Wire
+
+final class DateFormatterTests: XCTestCase {
+
+    override func setUp() {
+        super.setUp()
+    }
+    
+    override func tearDown() {
+        super.tearDown()
+    }
+
+    func testThatDateStringDoNotContainsYearIfDateIsToday(){
+        // GIVEN
+        let date = Date()
+        let dateFormatter = Date.localizedDateFormatter(date: date)
+        let dateString = dateFormatter.string(from: date)
+
+        let calendar = Calendar.current
+        let year = calendar.component(.year, from: date)
+
+        // WHEN & THEN
+        XCTAssertFalse(dateString.contains(String(year)))
+    }
+}
