--- WireExtensionComponents/Utilities/NSDate+Locale.swift
@@ -1,44 +1,0 @@
-//
-// Wire
-// Copyright (C) 2017 Wire Swiss GmbH
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-//
-
-
-import Foundation
-
-extension NSDate {
-
-    /// Create a NSDateFormatter depends on the date is in this year or not
-    ///
-    /// - Parameter date: date to create formatter
-    /// - Returns: a NSDateFormatter object. If the date param' s year is same as today, return a NSDateFormatter without year component, otherwise return a NSDateFormatter with year component.
-    @objc public static func localizedDateFormatter(date: NSDate) -> DateFormatter {
-        let today = Date()
-        let isThisYear = Calendar.current.isDate(date as Date, equalTo: today, toGranularity: .year)
-        let locale = NSLocale.current as NSLocale
-
-        var formatString: String?
-        if isThisYear {
-            formatString = DateFormatter.dateFormat(fromTemplate: "EEEEdMMMM", options: 0, locale: locale as Locale)
-        }
-        else {
-            formatString = DateFormatter.dateFormat(fromTemplate: "EEEEdMMMMYYYY", options: 0, locale: locale as Locale)
-        }
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = formatString
-        return dateFormatter
-    }
-}
